+++ src/components/molecules/TopAssetsChart/index.tsx
@@ -5,7 +5,7 @@ import { useEnvironment } from "src/context/EnvironmentContext";
 import { BREAKPOINTS } from "src/consts";
 import { BlockchainIcon } from "src/components/atoms";
 import { WormholeBrand } from "src/components/molecules";
-import { numberToSuffix } from "src/utils/number";
+import { formatterYAxisTopAssets } from "src/utils/apexChartUtils";
 import { AssetsByVolumeTransformed } from "src/api/guardian-network/types";
 import NoColorlessIcon from "src/icons/blockchains/colorless/noIcon.svg";
 import "./styles.scss";
@@ -26,46 +26,6 @@ const TopAssetsChart = ({ rowSelected, top7AssetsData, width }: Props) => {
   const isMobile = width < BREAKPOINTS.tablet;
   const isTabletOrMobile = width < BREAKPOINTS.desktop;
   const isDesktop = width >= BREAKPOINTS.desktop && width < BREAKPOINTS.bigDesktop;
-  /*  let allMillionsYIncludePointZero = false;
-  let allThousandsYIncludesPointZero = false; */
-
-  const formatter = (() => {
-    let allMillionsYIncludePointZero = false;
-    let allThousandsYIncludePointZero = false;
-
-    return function (vol: number, opts: any) {
-      let result = numberToSuffix(vol);
-      const allYAxis = opts?.w?.globals?.yAxisScale?.[0]?.result;
-
-      if (allYAxis && Array.isArray(allYAxis)) {
-        const allYAxisFormatted = allYAxis?.map((value: number) => {
-          return numberToSuffix(value);
-        });
-
-        allMillionsYIncludePointZero = allYAxisFormatted
-          .filter((value: string) => value.includes("M"))
-          .every((value: string) => value.includes(".0"));
-
-        allThousandsYIncludePointZero = allYAxisFormatted
-          .filter((value: string) => value.includes("K"))
-          .every((value: string) => value.includes(".0"));
-      }
-
-      if (vol >= 1000000 && allMillionsYIncludePointZero) {
-        result = result.replace(".0", "");
-      }
-
-      if (vol >= 1000 && vol < 1000000 && allThousandsYIncludePointZero) {
-        result = result.replace(".0", "");
-      }
-
-      if (vol < 1 && vol > 0) {
-        result = Number(result).toFixed(1);
-      }
-
-      return `$${result}`;
-    };
-  })();
 
   useEffect(() => {
     const getLabelsPos = () => {
@@ -279,7 +239,7 @@ const TopAssetsChart = ({ rowSelected, top7AssetsData, width }: Props) => {
             tickAmount: 8,
             axisTicks: { show: false },
             labels: {
-              formatter: formatter,
+              formatter: formatterYAxisTopAssets,
               minWidth: isMobile ? 48 : 64,
               maxWidth: isMobile ? 48 : 64,
               align: "right",
